Changed some files to test the expected value approach
- main.py: Change the ensemble of Rosenbrock and Rastrigin.
- test_resource_test_functions.h: Adding ensemble functions of
Rosenbrock and Rastrigin.
- test_tr-dfo_exp-value.cpp: Adding a function named Print_log() that
creates a file in the format .txt to save the log of iterations proceded
by Expected Value Class.
- TrustRegionOptimization class: Change of printIteration function to
save the points that are being used by the TR model and seting the
precision of datas.
- ensemble_exp_value.cpp: Disconsidering infinity cases in incubemt
solution provided by the optimizator.
- CMakeLists.txt: Commenting the test_tr-dfo.cpp because two different
files can not use a format .hpp file simutaniously.