Implemented new criticality logic: Criticality step may return false (e.g., b/c conditions not met), e.g., criticality step has not been performed even through criticality checks are enabled (gradient norm below epsC threshold). [Check btw that epsC is not too high compared to regular gradient norm.]
In such case, we should not reduce the radius at next radius update. Rather radius should be reduced
if criticality step has been successful, and kept the same or increased if we are not in a critical
neighborhood.

Bug: radius in critStep() code was not reverted to original size when stepping outside of computation
to evaluate cost function (after which we redo computation, i.e., re-apply the radius reduction within
this loop).